[C-736] Refactor AudiusLibs to typescript
