+++ libs/package.json
@@ -4,7 +4,7 @@
   "description": "",
   "main": "dist/index.cjs.js",
   "module": "dist/index.esm.js",
-  "types": "dist/types.d.ts",
+  "types": "dist/src/index.d.ts",
   "browser": {
     "./dist/index.cjs.js": "./dist/index.browser.cjs.js",
     "./dist/index.esm.js": "./dist/index.browser.esm.js"
@@ -119,10 +119,12 @@
     "rollup-plugin-ignore": "1.0.10",
     "rollup-plugin-polyfill-node": "0.9.0",
     "rollup-plugin-terser": "^7.0.2",
+    "rollup-plugin-typescript2": "^0.32.1",
     "sinon": "9.0.2",
     "standard": "16.0.4",
     "ts-mocha": "9.0.2",
     "ts-node": "10.7.0",
+    "type-fest": "^2.17.0",
     "typescript": "4.6.2"
   },
   "repository": {
