+++ libs/rollup.config.js
@@ -2,7 +2,7 @@ import commonjs from '@rollup/plugin-commonjs'
 import babel from '@rollup/plugin-babel'
 import json from '@rollup/plugin-json'
 import resolve from '@rollup/plugin-node-resolve'
-import typescript from '@rollup/plugin-typescript'
+import typescript from 'rollup-plugin-typescript2'
 import { terser } from 'rollup-plugin-terser'
 import dts from 'rollup-plugin-dts'
 import nodePolyfills from 'rollup-plugin-polyfill-node'
@@ -23,6 +23,8 @@ const external = [
 
 const commonConfig = {
   plugins: [
+    resolve({ extensions, preferBuiltins: true }),
+    typescript(),
     commonjs({
       extensions,
       dynamicRequireTargets: [
@@ -31,9 +33,7 @@ const commonConfig = {
       ]
     }),
     babel({ babelHelpers: 'bundled', extensions }),
-    json(),
-    resolve({ extensions, preferBuiltins: true }),
-    typescript()
+    json()
   ],
   external
 }
@@ -133,73 +133,72 @@ export default [
    * SDK
    */
   {
-    input: 'src/index.js',
+    input: 'src/index.ts',
     output: [
       { file: pkg.main, format: 'cjs', sourcemap: true },
       { file: pkg.module, format: 'es', sourcemap: true }
     ],
     ...commonConfig
   },
+  // {
+  //   input: 'dist-types/src/index.d.ts',
+  //   output: [{ file: pkg.types, format: 'es' }],
+  //   ...commonTypeConfig
+  // },
 
-  {
-    input: 'src/types.ts',
-    output: [{ file: pkg.types, format: 'es' }],
-    ...commonTypeConfig
-  },
+  // /**
+  //  * SDK bundled for a browser environment (includes polyfills for node libraries)
+  //  * Does not include libs but does include polyfills
+  //  */
+  // {
+  //   input: 'src/sdk/index.ts',
+  //   output: [
+  //     { file: 'dist/index.browser.cjs.js', format: 'es', sourcemap: true },
+  //     { file: 'dist/index.browser.esm.js', format: 'es', sourcemap: true }
+  //   ],
+  //   ...browserConfig
+  // },
 
-  /**
-   * SDK bundled for a browser environment (includes polyfills for node libraries)
-   * Does not include libs but does include polyfills
-   */
-  {
-    input: 'src/sdk/index.ts',
-    output: [
-      { file: 'dist/index.browser.cjs.js', format: 'es', sourcemap: true },
-      { file: 'dist/index.browser.esm.js', format: 'es', sourcemap: true }
-    ],
-    ...browserConfig
-  },
+  // /**
+  //  * SDK bundled for prebuilt package file to be used in browser
+  //  * Does not include libs but does include polyfills and all deps/dev deps
+  //  */
+  // {
+  //   input: 'src/sdk/sdkBrowserDist.ts',
+  //   output: [
+  //     {
+  //       file: 'dist/sdk.js',
+  //       globals: {
+  //         web3: 'window.Web3'
+  //       },
+  //       format: 'iife',
+  //       esModule: false,
+  //       sourcemap: true,
+  //       plugins: [terser()]
+  //     }
+  //   ],
+  //   ...browserDistFileConfig
+  // },
 
-  /**
-   * SDK bundled for prebuilt package file to be used in browser
-   * Does not include libs but does include polyfills and all deps/dev deps
-   */
-  {
-    input: 'src/sdk/sdkBrowserDist.ts',
-    output: [
-      {
-        file: 'dist/sdk.js',
-        globals: {
-          web3: 'window.Web3'
-        },
-        format: 'iife',
-        esModule: false,
-        sourcemap: true,
-        plugins: [terser()]
-      }
-    ],
-    ...browserDistFileConfig
-  },
+  // /**
+  //  * Legacy bundle for a browser environment
+  //  * Includes libs but does not include polyfills
+  //  */
+  // {
+  //   input: 'src/index.ts',
+  //   output: [{ file: 'dist/legacy.js', format: 'cjs', sourcemap: true }],
+  //   ...browserLegacyConfig
+  // },
 
-  /**
-   * Legacy bundle for a browser environment
-   * Includes libs but does not include polyfills
-   */
-  {
-    input: 'src/index.js',
-    output: [{ file: 'dist/legacy.js', format: 'cjs', sourcemap: true }],
-    ...browserLegacyConfig
-  },
+  // // {
+  // //   input: 'src/types.ts',
+  // //   output: [{ file: 'dist/legacy.d.ts', format: 'es' }],
+  // //   ...commonTypeConfig
+  // // },
 
-  {
-    input: 'src/types.ts',
-    output: [{ file: 'dist/legacy.d.ts', format: 'es' }],
-    ...commonTypeConfig
-  },
-
-  /**
-   * core (used for eager requests)
-   */
+  // /**
+  //  * core (used for eager requests)
+  //  */
   {
     input: 'src/core.ts',
     output: [
@@ -207,10 +206,10 @@ export default [
       { file: 'dist/core.esm.js', format: 'es', sourcemap: true }
     ],
     ...commonConfig
-  },
-  {
-    input: 'src/core.ts',
-    output: [{ file: pkg.coreTypes, format: 'es' }],
-    ...commonTypeConfig
   }
+  // {
+  //   input: 'src/core.ts',
+  //   output: [{ file: pkg.coreTypes, format: 'es' }],
+  //   ...commonTypeConfig
+  // }
 ]
