+++ libs/src/AudiusLibs.ts
@@ -1,77 +1,130 @@
-const packageJSON = require('../package.json')
+import type { provider } from 'web3-core'
+import type { Merge } from 'type-fest'
 
-const { EthWeb3Manager } = require('./services/ethWeb3Manager')
-const { SolanaAudiusData } = require('./services/solanaAudiusData/index')
-const { Web3Manager } = require('./services/web3Manager')
-const { EthContracts } = require('./services/ethContracts')
-const {
+import { EthWeb3Config, EthWeb3Manager } from './services/ethWeb3Manager'
+
+import packageJSON from '../package.json'
+import {
+  AnchorAudiusDataConfig,
+  SolanaAudiusData
+} from './services/solanaAudiusData'
+import { Web3Config, Web3Manager } from './services/web3Manager'
+import { EthContracts } from './services/ethContracts'
+import {
   SolanaWeb3Manager,
   SolanaUtils,
-  RewardsAttester
-} = require('./services/solana')
-const { AudiusContracts } = require('./services/dataContracts')
-const { IdentityService } = require('./services/identity')
-const { Comstock } = require('./services/comstock')
-const { Hedgehog } = require('./services/hedgehog')
-const { CreatorNode } = require('./services/creatorNode')
-const { DiscoveryProvider } = require('./services/discoveryProvider')
-const { Wormhole } = require('./services/wormhole')
-const { AudiusABIDecoder } = require('./services/ABIDecoder')
-const { SchemaValidator } = require('./services/schemaValidator')
-const { UserStateManager } = require('./userStateManager')
-const SanityChecks = require('./sanityChecks')
-const { Utils, Captcha } = require('./utils')
-const { ServiceProvider } = require('./api/ServiceProvider')
+  SolanaWeb3Config
+} from './services/solana'
+import { AudiusContracts } from './services/dataContracts'
+import { IdentityService } from './services/identity'
+import { Comstock } from './services/comstock'
+import { Hedgehog, HedgehogConfig } from './services/hedgehog'
+import type { Hedgehog as HedgehogBase } from '@audius/hedgehog'
+import {
+  CreatorNode,
+  CreatorNodeConfig,
+  MonitoringCallbacks
+} from './services/creatorNode'
+import {
+  DiscoveryProvider,
+  DiscoveryProviderConfig
+} from './services/discoveryProvider'
+import { Wormhole, WormholeConfig } from './services/wormhole'
+import { AudiusABIDecoder } from './services/ABIDecoder'
+import { Schemas, SchemaValidator } from './services/schemaValidator'
+import { UserStateManager } from './userStateManager'
+import { Utils, Captcha, Nullable, Logger, CaptchaConfig } from './utils'
+import { ServiceProvider } from './api/ServiceProvider'
 
-const { Account } = require('./api/Account')
-const { Users } = require('./api/Users')
-const { Track } = require('./api/Track')
-const { Playlists } = require('./api/Playlist')
-const { File } = require('./api/File')
-const { Rewards } = require('./api/Rewards')
-const { Reactions } = require('./api/Reactions')
-const Web3 = require('./web3')
+import { Account } from './api/Account'
+import { Users } from './api/Users'
+import { Track } from './api/Track'
+import { Playlists } from './api/Playlist'
+import { File } from './api/File'
+import { Rewards } from './api/Rewards'
+import { Reactions } from './api/Reactions'
+import Web3 from './web3'
 
-const { Keypair } = require('@solana/web3.js')
-const { PublicKey } = require('@solana/web3.js')
-const { getPlatformLocalStorage } = require('./utils/localStorage')
+import { Keypair, PublicKey } from '@solana/web3.js'
+import { getPlatformLocalStorage, LocalStorage } from './utils/localStorage'
 
-class AudiusLibs {
+type LibsIdentityServiceConfig = {
+  url: string
+  useHedgehogLocalStorage: boolean
+}
+
+type LibsHedgehogConfig = Omit<
+  HedgehogConfig,
+  'identityService' | 'localStorage'
+>
+
+type LibsSolanaWeb3Config = SolanaWeb3Config & {
+  // fee payer secret keys, if client wants to switch between different fee payers during relay
+  feePayerSecretKeys: Uint8Array[]
+}
+
+type LibsWormholeConfig = Merge<WormholeConfig, { rpcHosts: string | string[] }>
+
+type LibsDiscoveryProviderConfig = Omit<
+  DiscoveryProviderConfig,
+  'userStateManager' | 'ethContracts' | 'web3Manager'
+>
+
+type LibsComstockConfig = {
+  url: string
+}
+
+type AudiusLibsConfig = {
+  web3Config: Web3Config
+  ethWeb3Config: EthWeb3Config
+  solanaWeb3Config: SolanaWeb3Config
+  solanaAudiusDataConfig: AnchorAudiusDataConfig
+  identityServiceConfig: LibsIdentityServiceConfig
+  discoveryProviderConfig: LibsDiscoveryProviderConfig
+  creatorNodeConfig: CreatorNodeConfig
+  comstockConfig: LibsComstockConfig
+  wormholeConfig: WormholeConfig
+  captchaConfig: CaptchaConfig
+  hedgehogConfig: LibsHedgehogConfig
+  isServer: boolean
+  logger: Logger
+  isDebug: boolean
+  preferHigherPatchForPrimary: boolean
+  preferHigherPatchForSecondaries: boolean
+  localStorage: LocalStorage
+}
+
+export class AudiusLibs {
   /**
    * Configures an identity service wrapper
-   * @param {string} url
-   * @param {boolean?} useHedgehogLocalStorage whether or not to read hedgehog entropy in local storage
    */
-  static configIdentityService (url, useHedgehogLocalStorage = true) {
+  static configIdentityService(
+    url: string,
+    // whether or not to read hedgehog entropy in local storage
+    useHedgehogLocalStorage = true
+  ) {
     return { url, useHedgehogLocalStorage }
   }
 
   /**
    * Configures an identity service wrapper
-   * @param {string} url
    */
-  static configComstock (url) {
+  static configComstock(url: string) {
     return { url }
   }
 
-  /**
-   * Configures a creator node wrapper
-   * @param {string} fallbackUrl creator node endpoint to fall back to on requests
-   * @param {boolean} lazyConnect whether to delay connection to the node until the first
-   * request that requires a connection is made.
-   * @param {Set<string>?} passList whether or not to include only specified nodes (default null)
-   * @param {Set<string>?} blockList whether or not to exclude any nodes (default null)
-   * @param {object?} monitoringCallbacks callbacks to be invoked with metrics from requests sent to a service
-   * @param {function} monitoringCallbacks.request
-   * @param {function} monitoringCallbacks.healthCheck
-   * @param {boolean} writeQuorumEnabled whether or not to enforce waiting for replication to 2/3 nodes when writing data
-   */
-  static configCreatorNode (
-    fallbackUrl,
+  static configCreatorNode(
+    // creator node endpoint to fall back to on requests
+    fallbackUrl: string,
+    // whether to delay connection to the node until the first request that requires a connection is made.
     lazyConnect = false,
-    passList = null,
-    blockList = null,
-    monitoringCallbacks = {},
+    // whether or not to include only specified nodes (default null)
+    passList: Nullable<Set<string>> = null,
+    // whether or not to exclude any nodes (default null)
+    blockList: Nullable<Set<string>> = null,
+    // callbacks to be invoked with metrics from requests sent to a service
+    monitoringCallbacks: Nullable<MonitoringCallbacks> = {},
+    // whether or not to enforce waiting for replication to 2/3 nodes when writing data
     writeQuorumEnabled = false
   ) {
     return {
@@ -86,17 +139,15 @@ class AudiusLibs {
 
   /**
    * Configures an external web3 to use with Audius Libs (e.g. MetaMask)
-   * @param {string} registryAddress
-   * @param {Object} web3Provider equal to web.currentProvider
-   * @param {?number} networkId network chain id
-   * @param {?string} walletOverride wallet address to force use instead of the first wallet on the provided web3
-   * @param {?number} walletIndex if using a wallet returned from web3, pick the wallet at this index
    */
-  static async configExternalWeb3 (
-    registryAddress,
-    web3Provider,
-    networkId,
-    walletOverride = null
+  static async configExternalWeb3(
+    registryAddress: string,
+    // equal to web.currentProvider
+    web3Provider: string,
+    // network chain id
+    networkId: string,
+    // wallet address to force use instead of the first wallet on the provided web3
+    walletOverride: Nullable<string> = null
   ) {
     const web3Instance = await Utils.configureWeb3(web3Provider, networkId)
     if (!web3Instance) {
@@ -108,17 +159,19 @@ class AudiusLibs {
       useExternalWeb3: true,
       externalWeb3Config: {
         web3: web3Instance,
-        ownerWallet: walletOverride || wallets[0]
+        ownerWallet: walletOverride ?? wallets[0]
       }
     }
   }
 
   /**
    * Configures an internal web3 to use (via Hedgehog)
-   * @param {string} registryAddress
-   * @param {string | Web3 | Array<string>} providers web3 provider endpoint(s)
    */
-  static configInternalWeb3 (registryAddress, providers, privateKey) {
+  static configInternalWeb3(
+    registryAddress: string,
+    providers: provider,
+    privateKey: string
+  ) {
     let providerList
     if (typeof providers === 'string') {
       providerList = providers.split(',')
@@ -144,20 +197,16 @@ class AudiusLibs {
 
   /**
    * Configures an eth web3
-   * @param {string} tokenAddress
-   * @param {string} registryAddress
-   * @param {string | Web3 | Array<string>} providers web3 provider endpoint(s)
-   * @param {string?} ownerWallet optional owner wallet to establish who we are sending transactions on behalf of
-   * @param {string?} claimDistributionContractAddress
-   * @param {string?} wormholeContractAddress
    */
-  static configEthWeb3 (
-    tokenAddress,
-    registryAddress,
-    providers,
-    ownerWallet,
-    claimDistributionContractAddress,
-    wormholeContractAddress
+  static configEthWeb3(
+    tokenAddress: string,
+    registryAddress: string,
+    // web3 provider endpoint(s)
+    providers: provider,
+    // owner wallet to establish who we are sending transactions on behalf of
+    ownerWallet?: string,
+    claimDistributionContractAddress?: string,
+    wormholeContractAddress?: string
   ) {
     let providerList
     if (typeof providers === 'string') {
@@ -184,20 +233,14 @@ class AudiusLibs {
 
   /**
    * Configures wormhole
-   * @param {Object} config
-   * @param {string | Array<string>} config.rpcHosts
-   * @param {string} config.solBridgeAddress
-   * @param {string} config.solTokenBridgeAddress
-   * @param {string} config.ethBridgeAddress
-   * @param {string} config.ethTokenBridgeAddress
    */
-  static configWormhole ({
+  static configWormhole({
     rpcHosts,
     solBridgeAddress,
     solTokenBridgeAddress,
     ethBridgeAddress,
     ethTokenBridgeAddress
-  }) {
+  }: LibsWormholeConfig): WormholeConfig {
     let rpcHostList
     if (typeof rpcHosts === 'string') {
       rpcHostList = rpcHosts.split(',')
@@ -217,25 +260,8 @@ class AudiusLibs {
 
   /**
    * Configures a solana web3
-   * @param {Object} config
-   * @param {string} config.solanaClusterEndpoint the RPC endpoint to make requests against
-   * @param {string} config.mintAddress wAudio mint address
-   * @param {string} solanaTokenAddress native solana token program
-   * @param {string} claimableTokenPDA the generated program derived address we use so our
-   *  bank program can take ownership of accounts
-   * @param {string} feePayerAddress address for the fee payer for transactions
-   * @param {string} claimableTokenProgramAddress address of the audius user bank program
-   * @param {string} rewardsManagerProgramId address for the Rewards Manager program
-   * @param {string} rewardsManagerProgramPDA Rewards Manager PDA
-   * @param {string} rewardsManagerTokenPDA The PDA of the rewards manager funds holder account
-   * @param {boolean} useRelay Whether to use identity as a relay or submit transactions locally
-   * @param {Uint8Array} feePayerSecretKeys fee payer secret keys, if client wants to switch between different fee payers during relay
-   * @param {number} confirmationTimeout solana web3 connection confirmationTimeout in ms
-   * @param {PublicKey|string} audiusDataAdminStorageKeypairPublicKey admin storage PK for audius-data program
-   * @param {PublicKey|string} audiusDataProgramId program ID for the audius-data Anchor program
-   * @param {Idl} audiusDataIdl IDL for the audius-data Anchor program.
    */
-  static configSolanaWeb3 ({
+  static configSolanaWeb3({
     solanaClusterEndpoint,
     mintAddress,
     solanaTokenAddress,
@@ -251,7 +277,7 @@ class AudiusLibs {
     audiusDataAdminStorageKeypairPublicKey,
     audiusDataProgramId,
     audiusDataIdl
-  }) {
+  }: LibsSolanaWeb3Config): SolanaWeb3Config {
     if (audiusDataAdminStorageKeypairPublicKey instanceof String) {
       audiusDataAdminStorageKeypairPublicKey = new PublicKey(
         audiusDataAdminStorageKeypairPublicKey
@@ -271,9 +297,9 @@ class AudiusLibs {
       rewardsManagerProgramPDA,
       rewardsManagerTokenPDA,
       useRelay,
-      feePayerKeypairs: feePayerSecretKeys
-        ? feePayerSecretKeys.map((key) => Keypair.fromSecretKey(key))
-        : null,
+      feePayerKeypairs: feePayerSecretKeys.map((key) =>
+        Keypair.fromSecretKey(key)
+      ),
       confirmationTimeout,
       audiusDataAdminStorageKeypairPublicKey,
       audiusDataProgramId,
@@ -283,16 +309,67 @@ class AudiusLibs {
 
   /**
    * Configures a solana audius-data
-   * @param {Object} config
-   * @param {string} config.programId Program ID of the audius data program
-   * @param {string} config.adminAccount Public Key of admin account
    */
-  static configSolanaAudiusData ({ programId, adminAccount }) {
+  static configSolanaAudiusData({
+    programId,
+    adminAccount
+  }: AnchorAudiusDataConfig) {
     return {
       programId,
       adminAccount
     }
   }
+
+  version: string
+
+  ethWeb3Config: EthWeb3Config
+  web3Config: Web3Config
+  solanaWeb3Config: SolanaWeb3Config
+  solanaAudiusDataConfig: AnchorAudiusDataConfig
+  identityServiceConfig: LibsIdentityServiceConfig
+  creatorNodeConfig: CreatorNodeConfig
+  discoveryProviderConfig: LibsDiscoveryProviderConfig
+  comstockConfig: LibsComstockConfig
+  wormholeConfig: WormholeConfig
+  captchaConfig: CaptchaConfig
+  hedgehogConfig: LibsHedgehogConfig
+  isServer: boolean
+  isDebug: boolean
+  logger: Logger
+
+  AudiusABIDecoder: AudiusABIDecoder
+  Utils: Utils
+
+  // Services to initialize. Initialized in .init().
+  userStateManager: Nullable<UserStateManager>
+  identityService: Nullable<IdentityService>
+  hedgehog: Nullable<HedgehogBase>
+  discoveryProvider: Nullable<DiscoveryProvider>
+  ethWeb3Manager: Nullable<EthWeb3Manager>
+  ethContracts: Nullable<EthContracts>
+  web3Manager: Nullable<Web3Manager>
+  solanaWeb3Manager: Nullable<SolanaWeb3Manager>
+  solanaAudiusData: Nullable<SolanaAudiusData>
+  contracts: Nullable<AudiusContracts>
+  wormholeClient: Nullable<Wormhole>
+  creatorNode: Nullable<CreatorNode>
+  captcha: Nullable<Captcha>
+  schemas?: Schemas
+  comstock: Nullable<Comstock>
+
+  // API
+  ServiceProvider: Nullable<ServiceProvider>
+  Account: Nullable<Account>
+  User: Nullable<Users>
+  Track: Nullable<Track>
+  Playlist: Nullable<Playlists>
+  File: Nullable<File>
+  Rewards: Nullable<Rewards>
+  Reactions: Nullable<Reactions>
+
+  preferHigherPatchForPrimary: boolean
+  preferHigherPatchForSecondaries: boolean
+  localStorage: LocalStorage
 
   /**
    * Constructs an Audius Libs instance with configs.
@@ -304,7 +381,7 @@ class AudiusLibs {
    *  })
    *  await audius.init()
    */
-  constructor ({
+  constructor({
     web3Config,
     ethWeb3Config,
     solanaWeb3Config,
@@ -322,7 +399,7 @@ class AudiusLibs {
     preferHigherPatchForPrimary = true,
     preferHigherPatchForSecondaries = true,
     localStorage = getPlatformLocalStorage()
-  }) {
+  }: AudiusLibsConfig) {
     // set version
 
     this.version = packageJSON.version
@@ -354,11 +431,15 @@ class AudiusLibs {
     this.ethContracts = null
     this.web3Manager = null
     this.solanaWeb3Manager = null
-    this.anchorAudiusData = null
+    this.solanaAudiusData = null
+    this.wormholeClient = null
     this.contracts = null
     this.creatorNode = null
+    this.captcha = null
+    this.comstock = null
 
     // API
+    this.ServiceProvider = null
     this.Account = null
     this.User = null
     this.Track = null
@@ -378,7 +459,7 @@ class AudiusLibs {
   }
 
   /** Init services based on presence of a relevant config. */
-  async init () {
+  async init() {
     this.userStateManager = new UserStateManager({
       localStorage: this.localStorage
     })
@@ -409,14 +490,14 @@ class AudiusLibs {
     }
 
     /** Web3 Managers */
-    if (this.ethWeb3Config) {
+    if (this.ethWeb3Config && this.identityService && this.hedgehog) {
       this.ethWeb3Manager = new EthWeb3Manager({
         web3Config: this.ethWeb3Config,
         identityService: this.identityService,
         hedgehog: this.hedgehog
       })
     }
-    if (this.web3Config) {
+    if (this.web3Config && this.identityService && this.hedgehog) {
       this.web3Manager = new Web3Manager({
         web3Config: this.web3Config,
         identityService: this.identityService,
@@ -428,7 +509,7 @@ class AudiusLibs {
         this.identityService.setWeb3Manager(this.web3Manager)
       }
     }
-    if (this.solanaWeb3Config) {
+    if (this.solanaWeb3Config && this.identityService && this.web3Manager) {
       this.solanaWeb3Manager = new SolanaWeb3Manager(
         this.solanaWeb3Config,
         this.identityService,
@@ -436,7 +517,11 @@ class AudiusLibs {
       )
       await this.solanaWeb3Manager.init()
     }
-    if (this.solanaWeb3Manager && this.solanaAudiusDataConfig) {
+    if (
+      this.solanaWeb3Manager &&
+      this.solanaAudiusDataConfig &&
+      this.web3Manager
+    ) {
       this.solanaAudiusData = new SolanaAudiusData(
         this.solanaAudiusDataConfig,
         this.solanaWeb3Manager,
@@ -447,32 +532,31 @@ class AudiusLibs {
 
     /** Contracts - Eth and Data Contracts */
     const contractsToInit = []
-    if (this.ethWeb3Manager) {
+    if (this.ethWeb3Manager && this.ethWeb3Config) {
+      const {
+        tokenAddress,
+        registryAddress,
+        claimDistributionContractAddress,
+        wormholeContractAddress
+      } = this.ethWeb3Config
+
       this.ethContracts = new EthContracts({
         ethWeb3Manager: this.ethWeb3Manager,
-        tokenContractAddress: this.ethWeb3Config
-          ? this.ethWeb3Config.tokenAddress
-          : null,
-        registryAddress: this.ethWeb3Config
-          ? this.ethWeb3Config.registryAddress
-          : null,
-        claimDistributionContractAddress:
-          (this.ethWeb3Config &&
-            this.ethWeb3Config.claimDistributionContractAddress) ||
-          null,
-        wormholeContractAddress:
-          (this.ethWeb3Config && this.ethWeb3Config.wormholeContractAddress) ||
-          null,
+        tokenContractAddress: tokenAddress,
+        registryAddress,
+        claimDistributionContractAddress,
+        wormholeContractAddress,
         isServer: this.isServer,
         logger: this.logger,
         isDebug: this.isDebug
       })
+
       contractsToInit.push(this.ethContracts.init())
     }
-    if (this.web3Manager) {
+    if (this.web3Manager && this.web3Config) {
       this.contracts = new AudiusContracts(
         this.web3Manager,
-        this.web3Config ? this.web3Config.registryAddress : null,
+        this.web3Config.registryAddress,
         this.isServer,
         this.logger
       )
@@ -480,9 +564,11 @@ class AudiusLibs {
     }
     await Promise.all(contractsToInit)
     if (
+      this.hedgehog &&
       this.wormholeConfig &&
       this.ethWeb3Manager &&
       this.ethContracts &&
+      this.identityService &&
       this.solanaWeb3Manager
     ) {
       this.wormholeClient = new Wormhole(
@@ -491,17 +577,17 @@ class AudiusLibs {
         this.ethContracts,
         this.identityService,
         this.solanaWeb3Manager,
+        // @ts-expect-error -- typescript is weirdly incorrect here
         this.wormholeConfig.rpcHosts,
         this.wormholeConfig.solBridgeAddress,
         this.wormholeConfig.solTokenBridgeAddress,
         this.wormholeConfig.ethBridgeAddress,
-        this.wormholeConfig.ethTokenBridgeAddress,
-        this.isServer
+        this.wormholeConfig.ethTokenBridgeAddress
       )
     }
 
     /** Discovery Provider */
-    if (this.discoveryProviderConfig) {
+    if (this.discoveryProviderConfig && this.ethContracts && this.web3Manager) {
       this.discoveryProvider = new DiscoveryProvider({
         userStateManager: this.userStateManager,
         ethContracts: this.ethContracts,
@@ -513,10 +599,10 @@ class AudiusLibs {
     }
 
     /** Creator Node */
-    if (this.creatorNodeConfig) {
+    if (this.creatorNodeConfig && this.web3Manager && this.schemas) {
       const currentUser = this.userStateManager.getCurrentUser()
       const creatorNodeEndpoint = currentUser
-        ? CreatorNode.getPrimary(currentUser.creator_node_endpoint) ||
+        ? CreatorNode.getPrimary(currentUser.creator_node_endpoint) ??
           this.creatorNodeConfig.fallbackUrl
         : this.creatorNodeConfig.fallbackUrl
 
@@ -540,38 +626,57 @@ class AudiusLibs {
       this.comstock = new Comstock(this.comstockConfig.url)
     }
 
-    // Initialize apis
-    const services = [
-      this.userStateManager,
-      this.identityService,
-      this.hedgehog,
-      this.discoveryProvider,
-      this.web3Manager,
-      this.contracts,
-      this.ethWeb3Manager,
-      this.ethContracts,
-      this.solanaWeb3Manager,
-      this.anchorAudiusData,
-      this.wormholeClient,
-      this.creatorNode,
-      this.comstock,
-      this.captcha,
-      this.isServer,
-      this.logger
-    ]
-    this.ServiceProvider = new ServiceProvider(...services)
-    this.User = new Users(
-      this.ServiceProvider,
-      this.preferHigherPatchForPrimary,
-      this.preferHigherPatchForSecondaries,
-      ...services
-    )
-    this.Account = new Account(this.User, ...services)
-    this.Track = new Track(...services)
-    this.Playlist = new Playlists(...services)
-    this.File = new File(this.User, ...services)
-    this.Rewards = new Rewards(this.ServiceProvider, ...services)
-    this.Reactions = new Reactions(...services)
+    if (
+      this.userStateManager &&
+      this.identityService &&
+      this.hedgehog &&
+      this.discoveryProvider &&
+      this.discoveryProvider &&
+      this.web3Manager &&
+      this.contracts &&
+      this.ethWeb3Manager &&
+      this.ethContracts &&
+      this.solanaWeb3Manager &&
+      this.solanaAudiusData &&
+      this.wormholeClient &&
+      this.creatorNode &&
+      this.comstock &&
+      this.captcha
+    ) {
+      // Initialize apis
+      const services = [
+        this.userStateManager,
+        this.identityService,
+        this.hedgehog,
+        this.discoveryProvider,
+        this.web3Manager,
+        this.contracts,
+        this.ethWeb3Manager,
+        this.ethContracts,
+        this.solanaWeb3Manager,
+        this.solanaAudiusData,
+        this.wormholeClient,
+        this.creatorNode,
+        this.comstock,
+        this.captcha,
+        this.isServer,
+        this.logger
+      ] as const
+
+      this.ServiceProvider = new ServiceProvider(...services)
+      this.User = new Users(
+        this.ServiceProvider,
+        this.preferHigherPatchForPrimary,
+        this.preferHigherPatchForSecondaries,
+        ...services
+      )
+      this.Account = new Account(this.User, ...services)
+      this.Track = new Track(...services)
+      this.Playlist = new Playlists(...services)
+      this.File = new File(this.User, ...services)
+      this.Rewards = new Rewards(this.ServiceProvider, ...services)
+      this.Reactions = new Reactions(...services)
+    }
   }
 }
 
@@ -579,11 +684,7 @@ class AudiusLibs {
 // https://github.com/rollup/plugins/tree/master/packages/commonjs#defaultismoduleexports
 // exports.__esModule = true
 
-module.exports = AudiusLibs
+export { AudiusABIDecoder, Utils, SolanaUtils, CreatorNode }
 
-module.exports.AudiusABIDecoder = AudiusABIDecoder
-module.exports.Utils = Utils
-module.exports.SolanaUtils = SolanaUtils
-module.exports.SanityChecks = SanityChecks
-module.exports.RewardsAttester = RewardsAttester
-module.exports.CreatorNode = CreatorNode
+export { SanityChecks } from './sanityChecks'
+export { RewardsAttester } from './services/solana'
