--- libs/src/index.js
@@ -1,6 +1,0 @@
-module.exports.sdk = require('./sdk')
-
-/**
- * @deprecated Please migrate to using `sdk` when possible
- */
-module.exports.libs = require('./libs')
