+++ libs/src/index.ts
@@ -1,0 +1,6 @@
+export { sdk } from './sdk'
+
+/**
+ * @deprecated Please migrate to using `sdk` when possible
+ */
+export { AudiusLibs as libs } from './AudiusLibs'
