--- libs/src/libs.d.ts
@@ -1,3 +1,0 @@
-declare const libs: {}
-
-export default libs
