+++ libs/src/sanityChecks/index.d.ts
@@ -1,9 +1,7 @@
-type Options = { skipRollover: boolean }
-
 export class SanityChecks {
   libs: any
   options: Options
 
-  constructor(libsInstance: any, options?: Options): void
+  constructor(libsInstance: any, options?: any): void
   async run(creatorNodeWhitelist: string[] | null): Promise<void>
 }
