+++ libs/src/sanityChecks/index.js
@@ -6,7 +6,7 @@ const recoveryEmail = require('./needsRecoveryEmail')
 const assignReplicaSetIfNecessary = require('./assignReplicaSetIfNecessary')
 
 // Checks to run at startup to ensure a user is in a good state.
-class SanityChecks {
+export class SanityChecks {
   constructor (libsInstance, options = { skipRollover: false }) {
     this.libs = libsInstance
     this.options = options
@@ -21,9 +21,7 @@ class SanityChecks {
     await addSecondaries(this.libs)
     await assignReplicaSetIfNecessary(this.libs)
     await syncNodes(this.libs)
-    if (!this.options.skipRollover) await rolloverNodes(this.libs, creatorNodeWhitelist)
+    if (!this.options.skipRollover) { await rolloverNodes(this.libs, creatorNodeWhitelist) }
     await recoveryEmail(this.libs)
   }
 }
-
-module.exports = SanityChecks
