+++ libs/src/services/web3Manager/Web3Config.ts
@@ -2,6 +2,7 @@ import type Web3 from 'web3'
 import type Wallet from 'ethereumjs-wallet'
 
 export type Web3Config = {
+  registryAddress: string
   useExternalWeb3: boolean
   internalWeb3Config: {
     web3ProviderEndpoints: string[]
