--- libs/src/types.ts
@@ -1,10 +1,0 @@
-import AudiusLibs from './libs'
-
-export { sdk } from './sdk'
-export * as full from './sdk/api/generated/full'
-export * from './sdk/api/generated/default'
-
-export const libs: any = AudiusLibs
-export { Utils } from './utils'
-export { CreatorNode } from './services/creatorNode'
-export * from './sanityChecks'
