+++ libs/tsconfig.json
@@ -1,11 +1,16 @@
 {
   "extends": "@tsconfig/node16-strictest/tsconfig.json",
   "compilerOptions": {
-    "lib": ["es2021", "dom"],
+    "lib": ["es2022", "dom"],
+    "module": "ESNext",
+    "moduleResolution": "node",
     "checkJs": false,
     "resolveJsonModule": true,
     "typeRoots": ["./types", "./node_modules/@types"],
-    "exactOptionalPropertyTypes": false
+    "exactOptionalPropertyTypes": false,
+    "declaration": true,
+    "declarationDir": "dist-types"
   },
-  "exclude": ["**/*+(test).ts"]
+  "exclude": ["**/*+(test).ts", "dist", "dist-types"],
+  "include": ["src/**/*", "data-contracts", "types/**/*"]
 }
